+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features & enhancements 🚀
 
+* Editor | Opening a file by default app on device instead of downloading it when clicked (#1904)
 * Editor | Latex block in experimental mode. Based on WebView. (#1814)
 * Objects | Menu | Select a layout, which best suits your object: Note, Basic, Profile, or Task (#1894)  
 * Sets | Delete a relation from your set (#1890)